style: minor carousel restyling tweaks and add in missing sponsors
