+++ src/App.tsx
@@ -70,6 +70,7 @@ function App() {
                                                 const newPreferences = { ...cardViewPreferences, [id]: preference };
                                                 setCardViewPreferences(newPreferences);
                                             }}
+                                            now={now}
                                         />
                                     }
                                 />
