--- src/components/SponsorCarousel.css
@@ -1,146 +1,0 @@
-.carousel {
-    position: relative;
-    overflow: hidden;
-}
-
-.carousel__track {
-    display: flex;
-    align-items: center;
-    gap: 40px;
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    width: max-content;
-    animation: carousel 90s linear infinite;
-}
-
-@media (max-width: 600px) {
-    .carousel__track {
-        animation: carousel 120s linear infinite;
-    }
-    .carousel__image {
-        height: 25px;
-    }
-}
-
-.sponsors--dark .carousel__track {
-    background-color: var(--button-bg--hover);
-}
-
-.sponsors--light .carousel__track {
-    background-color: #ababab;
-}
-
-.carousel__item {
-    flex: 0 0 auto;
-}
-
-.carousel__image {
-    height: 50px;
-    width: auto;
-    object-fit: contain;
-    display: block;
-    margin: 20px 0px;
-    padding: 5px;
-}
-
-.sponsors--dark .carousel__image {
-    filter: brightness(0) invert(1);
-    opacity: 30%;
-}
-
-.sponsors--light .carousel__image {
-    filter: brightness(0) invert(0);
-    opacity: 50%;
-}
-
-@keyframes carousel {
-    from {
-        transform: translateX(0);
-    }
-    to {
-        transform: translateX(-50%);
-    }
-}
-
-.footer__sponsors-tab {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-}
-
-.sponsors--dark .footer__sponsors-tab {
-    background-color: var(--footer-bg);
-}
-
-.sponsors--light .footer__sponsors-tab {
-    background-color: #f9fafb;
-}
-
-.footer__sponsors-desc {
-    display: inline-flex;
-    align-items: center;
-    padding: 0px 14px;
-    border-radius: 10px 10px 0px 0px;
-    line-height: 1;
-    white-space: nowrap;
-    margin-left: 20px;
-}
-
-.sponsors--dark .footer__sponsors-desc {
-    background: var(--button-bg--hover);
-    border-bottom: 1px solid rgba(255, 255, 255, 0.08);
-}
-
-.sponsors--light .footer__sponsors-desc {
-    background: #ababab;
-    border-bottom: 1px solid rgb(136, 136, 136);
-}
-
-.sponsors--dark .footer__sponsors-text {
-    color: rgba(255, 255, 255, 0.7);
-}
-
-.sponsors--light .footer__sponsors-text {
-    color: rgba(0, 0, 0, 0.5);
-}
-
-.footer__sponsors-text {
-    all: unset;
-    padding-top: 14px;
-    padding-bottom: 14px;
-    border-radius: 10px 10px 0px 0px;
-    line-height: 1;
-    white-space: nowrap;
-}
-
-.footer__sponsors-text:hover {
-    cursor: pointer;
-}
-
-.footer__sponsors {
-    overflow: hidden;
-    max-height: 0;
-    opacity: 0;
-    transform: translateY(-8px);
-    transition:
-        max-height 0.5s cubic-bezier(0.25, 0.8, 0.25, 1),
-        opacity 0.3s ease-out,
-        transform 0.5s cubic-bezier(0.25, 0.8, 0.25, 1);
-}
-
-.footer__sponsors--open {
-    max-height: 200px;
-    opacity: 1;
-    transform: translateY(0);
-}
-
-.footer__sponsors-arrow {
-    display: inline-block;
-    margin-left: 8px;
-    transition: transform 0.3s ease;
-}
-
-.footer__sponsors-arrow--open {
-    transform: rotate(180deg);
-}
