+++ src/components/SponsorCarousel.tsx
@@ -1,6 +1,8 @@
 import { useState } from 'react';
-import './SponsorCarousel.css';
-import dropdownArrow from '../assets/control_button/dropdown_arrow.svg';
+import clsx from 'clsx';
+import css from './SponsorCarousel.module.css';
+import DropdownArrow from '../assets/control_button/dropdown_arrow.svg?react';
+import ScottyDog from '../assets/scottydog.svg?react';
 
 function SponsorCarousel({ darkMode, openByDefault }: { darkMode: boolean; openByDefault: boolean }) {
     const [carouselVisible, setCarouselVisible] = useState(openByDefault);
@@ -18,44 +20,29 @@ function SponsorCarousel({ darkMode, openByDefault }: { darkMode: boolean; openB
         };
     });
 
-    // double it because it loops seamlessly
-    const doubleLogos = [...logos, ...logos];
+    // double it to make it loop seamlessly
+    const doubleLogos = [...logos, ...logos.map((l) => ({ ...l, alt: `${l.alt}2` }))];
 
     return (
-        <div className={`${darkMode ? 'sponsors--dark' : 'sponsors--light'}`}>
-            <div className="footer__sponsors-tab">
-                <span className="footer__sponsors-desc">
-                    <button
-                        onClick={() => setCarouselVisible((v) => !v)}
-                        className="footer__sponsors-text"
-                        aria-expanded={carouselVisible}
-                    >
-                        ScottyLabs is sponsored by{' '}
-                        <span
-                            className={`footer__sponsors-arrow ${carouselVisible ? 'footer__sponsors-arrow--open' : ''}`}
-                        >
-                            <img src={dropdownArrow} alt="Dropdown arrow" />
-                        </span>
-                    </button>
-                </span>
-            </div>
-
-            <div
-                className={`footer__sponsors ${
-                    carouselVisible ? 'footer__sponsors--open' : 'footer__sponsors--closed'
-                }`}
-                aria-hidden={!carouselVisible}
-            >
-                <div className="footer__sponsors-carousel">
-                    <div className="carousel">
-                        <ul className="carousel__track">
-                            {doubleLogos.map((logo) => (
-                                <li className="carousel__item">
-                                    <img src={logo.src} alt="" className="carousel__image" />
-                                </li>
-                            ))}
-                        </ul>
-                    </div>
+        <div className={clsx(css.sponsors, darkMode ? css['sponsors--dark'] : css['sponsors--light'])}>
+            <div className={css[`carousel-container`]} aria-hidden={!carouselVisible}>
+                <button
+                    onClick={() => setCarouselVisible((v) => !v)}
+                    className={css['sponsors-button']}
+                    aria-expanded={carouselVisible}
+                    type="button"
+                >
+                    <ScottyDog className={css['sponsors-button__dog']} /> ScottyLabs is sponsored by{' '}
+                    <DropdownArrow className={css[`sponsors-button__arrow`]} />
+                </button>
+                <div className={css.carousel}>
+                    <ul className={css.carousel__track}>
+                        {doubleLogos.map((logo) => (
+                            <li className={css.carousel__item} key={logo.alt}>
+                                <img src={logo.src} alt={logo.alt} className={css.carousel__image} />
+                            </li>
+                        ))}
+                    </ul>
                 </div>
             </div>
         </div>
