+++ src/pages/ListPage.css
@@ -173,22 +173,3 @@
         fade-in 1s cubic-bezier(0.08, 0.67, 0.64, 1.01) 1s forwards,
         oscillate-opacity 2s ease-in-out var(--oscillate-delay) infinite;
 }
-
-.footer {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.footer__text {
-    flex: 1;
-}
-
-.footer__logo-mobile {
-    display: none;
-}
-
-.footer__logo-desktop {
-    display: block;
-}
