+++ src/pages/ListPage.tsx
@@ -1,61 +1,37 @@
-import { Typography, Alert, styled } from '@mui/material';
+import { Alert, styled } from '@mui/material';
 import { useEffect, useLayoutEffect, useMemo, useReducer, useRef, useState } from 'react';
+import { DateTime } from 'luxon';
+
 import { getGreetings } from '../util/greeting';
 import './ListPage.css';
 import { IReadOnlyLocation_Combined } from '../types/locationTypes';
-
 import SelectLocation from '../components/SelectLocation';
 import SearchBar from '../components/SearchBar';
-import { useTheme } from '../ThemeProvider';
+import { useThemeContext } from '../ThemeProvider';
 import IS_MIKU_DAY from '../util/constants';
 import mikuKeychainUrl from '../assets/miku/miku-keychain.svg';
-import footerMikuUrl from '../assets/miku/miku2.png';
 import mikuBgUrl from '../assets/miku/miku.jpg';
 import EateryCardGrid from './EateryCardGrid';
 import useFilteredLocations from './useFilteredLocations';
-import env from '../env';
 import { CardViewPreference } from '../util/storage';
-import SponsorCarousel from '../components/SponsorCarousel';
-
-const LogoText = styled(Typography)({
-    color: 'var(--logo-first-half)',
-    padding: 0,
-    fontFamily: 'var(--text-primary-font)',
-    fontWeight: 800,
-});
-
-const FooterText = styled(Typography)({
-    color: 'var(--text-primary)',
-    marginBottom: 20,
-    fontSize: 16,
-});
+import Footer from '../components/Footer';
 
 const StyledAlert = styled(Alert)({
     backgroundColor: 'var(--main-bg-accent)',
     color: 'var(--text-primary)',
 });
 
-function getPittsburghTime() {
-    const now = new Date();
-    const options: Intl.DateTimeFormatOptions = {
-        timeZone: 'America/New_York',
-        hour: 'numeric',
-        minute: 'numeric',
-        second: 'numeric',
-        timeZoneName: 'short',
-    };
-    return now.toLocaleString('en-US', options);
-}
-
 function ListPage({
     locations,
     updateCardViewPreference,
+    now,
 }: {
     locations: IReadOnlyLocation_Combined[] | undefined;
+    now: DateTime;
     updateCardViewPreference: (id: string, newStatus: CardViewPreference) => void;
 }) {
-    const { theme, updateTheme } = useTheme();
     const shouldAnimateCards = useRef(true);
+    const { theme, updateTheme } = useThemeContext();
 
     // permanently cut out animation when user filters cards,
     // so we don't end up with some cards (but not others)
@@ -72,7 +48,6 @@ function ListPage({
         },
         '',
     );
-    const [emails, setEmails] = useState<{ name: string; email: string }[]>([]);
     const [showOfflineAlert, setShowOfflineAlert] = useState(!navigator.onLine);
 
     const { mobileGreeting, desktopGreeting } = useMemo(
@@ -85,20 +60,6 @@ function ListPage({
         searchQuery,
         locationFilterQuery,
     });
-
-    // Fetch emails on mount
-    useEffect(() => {
-        async function fetchEmails() {
-            try {
-                const res = await fetch(`${env.VITE_API_URL}/api/emails`);
-                const json = await res.json();
-                setEmails(json);
-            } catch (err) {
-                console.error('Failed to fetch emails:', err);
-            }
-        }
-        fetchEmails();
-    }, []);
 
     // Load query from URL
     useLayoutEffect(() => {
@@ -176,94 +137,7 @@ function ListPage({
                     }}
                 />
             </div>
-
-            <footer className="footer">
-                <div className="footer__text">
-                    {theme === 'miku' ? (
-                        <FooterText>
-                            Blue hair, blue tie, hiding in your wifi
-                            <br />
-                            All times are displayed in Pittsburgh local time ({getPittsburghTime()}).
-                        </FooterText>
-                    ) : (
-                        <>
-                            <FooterText>
-                                All times are displayed in Pittsburgh local time ({getPittsburghTime()}).
-                            </FooterText>
-                            <FooterText>
-                                If you encounter any problems, please fill out our{' '}
-                                <a href="https://forms.gle/7JxgdgDhWMznQJdk9" style={{ color: 'white' }}>
-                                    feedback form
-                                </a>{' '}
-                                (the fastest way to reach us!).
-                            </FooterText>
-                            <FooterText>
-                                Otherwise, reach out to{' '}
-                                {emails.length > 0 ? (
-                                    emails.map((person, idx) => (
-                                        <span key={person.email}>
-                                            <a href={`mailto:${person.email}`} style={{ color: 'white' }}>
-                                                {person.name}
-                                            </a>
-                                            {idx < emails.length - 2 ? ', ' : ''}
-                                            {/* eslint-disable-next-line no-nested-ternary */}
-                                            {idx === emails.length - 2 ? (emails.length > 2 ? ', or ' : ' or ') : ''}
-                                        </span>
-                                    ))
-                                ) : (
-                                    <span>
-                                        <a href="mailto:hello@scottylabs.org" style={{ color: 'white' }}>
-                                            ScottyLabs
-                                        </a>
-                                    </span>
-                                )}
-                                .
-                            </FooterText>
-                            <FooterText>
-                                To provide feedback on your dining experience, please contact{' '}
-                                <a href="mailto:dining@andrew.cmu.edu" style={{ color: 'white' }}>
-                                    Dining Services
-                                </a>{' '}
-                                or take the{' '}
-                                <a href="https://forms.gle/fTnWrS7jkTFRB14DA" style={{ color: 'white' }}>
-                                    dining survey
-                                </a>
-                                .
-                            </FooterText>
-                            <FooterText>
-                                Made with ❤️ by the{' '}
-                                <a href="https://scottylabs.org" style={{ color: 'white' }}>
-                                    ScottyLabs
-                                </a>{' '}
-                                Tech Committee (not the official{' '}
-                                <a
-                                    href="https://apps.studentaffairs.cmu.edu/dining/conceptinfo/Schedule"
-                                    target="_blank"
-                                    rel="noreferrer"
-                                    style={{ color: 'white' }}
-                                >
-                                    dining website
-                                </a>
-                                ).
-                            </FooterText>
-                        </>
-                    )}
-                    <div className="footer__logo-desktop">
-                        <LogoText variant="h4">
-                            cmu
-                            <span style={{ color: 'var(--logo-second-half)' }}>:eats</span>
-                        </LogoText>
-                    </div>
-                </div>
-                <div className="footer__logo-mobile">
-                    <LogoText variant="h4">
-                        cmu
-                        <span style={{ color: 'var(--logo-second-half)' }}>:eats</span>
-                    </LogoText>
-                </div>
-                {theme === 'miku' && <img src={footerMikuUrl} alt="miku!" className="footer__miku" />}
-            </footer>
-            <SponsorCarousel darkMode openByDefault />
+            <Footer now={now} />
             <link rel="prefetch" href={mikuBgUrl} />
         </div>
     );
